requirements.txt
fastapi
uvicorn
python-multipart
jinja2
numpy
scipy
Pillow
pytest
matplotlib
httpx